# Optimize `getPartitionEndRow` for window function