--- a/dbms/src/DataStreams/WindowBlockInputStream.cpp
+++ b/dbms/src/DataStreams/WindowBlockInputStream.cpp
@@ -124,13 +124,13 @@ Block WindowBlockInputStream::readImpl()
 // Judge whether current_partition_row is end row of partition in current block
 bool WindowTransformAction::isDifferentFromPrevPartition(UInt64 current_partition_row)
 {
-    const auto reference_columns = inputAt(prev_frame_start);
-    const auto compared_columns = inputAt(partition_end);
+    const Columns & reference_columns = inputAt(prev_frame_start);
+    const Columns & compared_columns = inputAt(partition_end);
 
     for (size_t i = 0; i < partition_column_indices.size(); ++i)
     {
-        const auto reference_column = reference_columns[partition_column_indices[i]];
-        const auto * compared_column = compared_columns[partition_column_indices[i]].get();
+        const ColumnPtr & reference_column = reference_columns[partition_column_indices[i]];
+        const ColumnPtr & compared_column = compared_columns[partition_column_indices[i]];
         if (window_description.partition_by[i].collator)
         {
             if (compared_column->compareAt(current_partition_row,
@@ -240,7 +240,9 @@ Int64 WindowTransformAction::getPartitionEndRow(size_t block_rows)
     Int64 left = partition_end.row;
     Int64 right = block_rows - 1;
 
-    for (; left <= std::min(left + 9, right); ++left)
+    // Compare several times first.
+    // It will benefit if the partition end is very close.
+    for (; left <= std::min(left + 2, right); ++left)
     {
         if (isDifferentFromPrevPartition(left))
         {
@@ -248,7 +250,7 @@ Int64 WindowTransformAction::getPartitionEndRow(size_t block_rows)
         }
     }
 
-    while (left <= right)
+    while (left < right)
     {
         Int64 middle = (left + right) << 1;
         if (isDifferentFromPrevPartition(middle))